# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Snow School Scheduler API
  description: スキー・スノーボードスクールのシフト管理システム API
  version: 1.0.0
  contact:
    name: Snow School Scheduler
servers:
  - url: http://localhost:3000
    description: 開発環境
  - url: https://snow-school-scheduler-api.example.com
    description: 本番環境
paths:
...
  /api/shifts/prepare:
    get:
      summary: シフト準備（重複チェック・編集準備）
      description: 指定された日付・部門・シフト種類の組み合わせで既存シフトの重複チェックと編集準備を行う統合エンドポイント
      tags:
        - シフト管理
      parameters:
        - name: date
          in: query
          required: true
          description: シフト日付（YYYY-MM-DD形式）
          schema:
            type: string
            format: date
            example: '2024-01-15'
        - name: departmentId
          in: query
          required: true
          description: 部門ID
          schema:
            type: integer
            example: 1
        - name: shiftTypeId
          in: query
          required: true
          description: シフト種類ID
          schema:
            type: integer
            example: 2
      responses:
        '200':
          description: シフト準備完了（新規作成モードまたは編集モード）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrepareShiftResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/shifts/[id]:
    parameters:
      - $ref: '#/components/parameters/Id'

    get:
      summary: シフト詳細取得
      description: 指定されたシフトの詳細情報を取得（割り当て統計付き）
      tags:
        - シフト管理
      responses:
        '200':
          description: シフト詳細
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShiftWithStats'
                      message:
                        type: string
                        example: 'Shift operation completed successfully'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: リソースのID
      schema:
        type: integer
  responses:
    NotFound:
      description: リソースが見つかりません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Resource not found'
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Validation failed'
    ServerError:
      description: サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Internal server error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InstructorSummary:
      type: object
      properties:
        id:
          type: integer
        lastName:
          type: string
          example: '山田'
        firstName:
          type: string
          example: '太郎'
        status:
          type: string
          enum: [ACTIVE, INACTIVE, RETIRED]
          default: ACTIVE
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
          example: null
...
    ApiErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Invalid invitation token'

    # 部門関連
    Department:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
          example: 'ski'
        name:
          type: string
          example: 'スキー'
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ShiftType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: 'レッスン'
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        departmentId:
          type: integer
        shiftTypeId:
          type: integer
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        department:
          $ref: '#/components/schemas/Department'
        shiftType:
          $ref: '#/components/schemas/ShiftType'
...
    ShiftWithStats:
      allOf:
        - $ref: '#/components/schemas/Shift'
        - type: object
          properties:
            assignments:
              type: array
              items:
                $ref: '#/components/schemas/ShiftAssignment'
            assignedCount:
              type: integer
              description: 現在割り当てられているインストラクター数

    # シフト割り当て関連
    ShiftAssignment:
      type: object
      properties:
        id:
          type: integer
        shiftId:
          type: integer
        instructorId:
          type: integer
        assignedAt:
          type: string
          format: date-time
        instructor:
          $ref: '#/components/schemas/InstructorSummary'
...
    PrepareShiftResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            mode:
              type: string
              enum: ['create', 'edit']
              description: '作成モードまたは編集モード'
              example: 'edit'
            shift:
              allOf:
                - $ref: '#/components/schemas/ExistingShiftData'
                - type: object
                  nullable: true
                  description: '編集モードの場合の既存シフトデータ（新規作成モードではnull）'
            formData:
              $ref: '#/components/schemas/PrepareShiftFormData'
        error:
          type: string
          nullable: true
          example: null
    ExistingShiftData:
      type: object
      description: '既存シフトの詳細データ（編集モード用）'
      properties:
        id:
          type: integer
          example: 1
        date:
          type: string
          format: date
          example: '2024-01-15'
        departmentId:
          type: integer
          example: 1
        shiftTypeId:
          type: integer
          example: 2
        description:
          type: string
          nullable: true
          example: 'モーニングレッスン'
        department:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: 'スキー'
            code:
              type: string
              example: 'ski'
        shiftType:
          type: object
          properties:
            id:
              type: integer
              example: 2
            name:
              type: string
              example: 'レッスン'
        assignments:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                example: 1
              instructor:
                type: object
                properties:
                  id:
                    type: integer
                    example: 3
                  lastName:
                    type: string
                    example: '山田'
                  firstName:
                    type: string
                    example: '太郎'
        assignedCount:
          type: integer
          description: '現在割り当てられているインストラクター数'
          example: 2
    PrepareShiftFormData:
      type: object
      description: 'フォーム事前設定用のデータ'
      properties:
        date:
          type: string
          format: date
          example: '2024-01-15'
        departmentId:
          type: integer
          example: 1
        shiftTypeId:
          type: integer
          example: 2
        description:
          type: string
          nullable: true
          example: 'モーニングレッスン'
        selectedInstructorIds:
          type: array
          items:
            type: integer
          description: '事前選択されるインストラクターIDの配列'
          example: [3, 5]
...